requirements.txt
numpy
matplotlib
pandas
ipykernel
openneuro-py
mne_bids